fixed apm2beta build with AP_Param

+++ ArduCopter/GCS_Mavlink.pde
@@ -1415,13 +1415,13 @@ void GCS_MAVLINK::handleMessage(mavlink_message_t* msg)
 					((AP_Int32 *)vp)->set_and_save(packet.param_value+rounding_addition);
                 } else if (var_type == AP_PARAM_INT16) {
 #if LOGGING_ENABLED == ENABLED
-					Log_Write_Data(3, ((AP_Int16 *)vp)->get());
+					Log_Write_Data(3, (int32_t)((AP_Int16 *)vp)->get());
 #endif
 					if (packet.param_value < 0) rounding_addition = -rounding_addition;
 					((AP_Int16 *)vp)->set_and_save(packet.param_value+rounding_addition);
                 } else if (var_type == AP_PARAM_INT8) {
 #if LOGGING_ENABLED == ENABLED
-					Log_Write_Data(4, ((AP_Int8 *)vp)->get());
+					Log_Write_Data(4, (int32_t)((AP_Int8 *)vp)->get());
 #endif
 					if (packet.param_value < 0) rounding_addition = -rounding_addition;
 					((AP_Int8 *)vp)->set_and_save(packet.param_value+rounding_addition);
